+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [4.1.0] - unreleased
 
 ### Added
+- Added new admin function: create, edit, delete, rename templates
 - Added Controller::is_online that replaces utils::is_online
 - Inline documentation for \DB\DB